+++ src/diffusers/loaders/__init__.py
@@ -9,13 +9,7 @@ _import_structure = {}
 if is_torch_available():
     _import_structure["single_file"] = ["FromOriginalControlnetMixin", "FromOriginalVAEMixin"]
     _import_structure["unet"] = ["UNet2DConditionLoadersMixin"]
-    _import_structure["utils"] = [
-        "AttnProcsLayers",
-        "PatchedLoraProjection",
-        "text_encoder_attn_modules",
-        "text_encoder_lora_state_dict",
-        "text_encoder_mlp_modules",
-    ]
+    _import_structure["utils"] = ["AttnProcsLayers"]
 
     if is_transformers_available():
         _import_structure["single_file"].extend(["FromSingleFileMixin"])
@@ -27,13 +21,7 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
     if is_torch_available():
         from .single_file import FromOriginalControlnetMixin, FromOriginalVAEMixin
         from .unet import UNet2DConditionLoadersMixin
-        from .utils import (
-            AttnProcsLayers,
-            PatchedLoraProjection,
-            text_encoder_attn_modules,
-            text_encoder_lora_state_dict,
-            text_encoder_mlp_modules,
-        )
+        from .utils import AttnProcsLayers
 
         if is_transformers_available():
             from .lora import LoraLoaderMixin, StableDiffusionXLLoraLoaderMixin
